feat(preview): #5 Phase 2-4 — complete xterm.js migration, retire AnsiUp pipeline
Drops the dispatcher, the legacy AnsiUp+Worker preview, and the
supporting ANSI-conversion helpers. PreviewPanel now is the xterm.js
implementation that landed behind the `tmai:preview-xterm` flag in
Phase 1 (#604) — Phase 1 served as the dogfood scaffold; this PR
completes the migration so production users see xterm by default.

### Removed

- `PreviewPanelLegacy.tsx` (the AnsiUp+Worker+innerHTML pipeline)
- `preview-content.ts` + tests (capHistoryLines / shrinkContentToWidth /
  trimPreviewContent / charColumns / splitPreviewContent — only the
  legacy panel called any of these)
- `lib/ansi-worker.ts` (the off-thread AnsiUp converter)
- `PreviewPanel-dispatcher.test.tsx` (dispatcher gone)
- `ansi_up` and `dompurify` package dependencies

### Renamed / refactored

- `PreviewPanelXterm.tsx` → `PreviewPanel.tsx` (single canonical
  implementation; the dispatcher shim is gone)
- xterm container gains `role="log"` + `aria-label` so existing
  interaction tests continue to locate it via `getByRole("log")`

### IME path

The hidden IME input + composition handler is intentionally retained.
xterm v5's IME handling still drops mid-composition keystrokes on the
Live preview surface; the hidden input is the proven path. A future
PR can revisit if and when xterm-native IME stabilises.

### Test plan

- Both `PreviewPanel-interaction.test.tsx` and `PreviewPanel-queue.test.tsx`
  now stub `useTerminal` so xterm's matchMedia / canvas calls are
  short-circuited under jsdom.
- `pnpm vitest run` — 41 files, 320 passed + 2 skipped (down from 43 /
  338: 2 deleted tests files for the removed dispatcher and
  preview-content helpers).
- `pnpm tsc --noEmit` — clean.
- `pnpm biome check src/` — no new diagnostics; the 6 warnings are
  pre-existing.

Refs: #5
