+++ clients/react/package.json
@@ -21,9 +21,7 @@
     "@xterm/addon-fit": "^0.11.0",
     "@xterm/xterm": "^6.0.0",
     "@xyflow/react": "^12.10.2",
-    "ansi_up": "^6.0.6",
     "clsx": "^2",
-    "dompurify": "^3.4.1",
     "lucide-react": "^1.11",
     "mermaid": "^11.14.0",
     "react": "^19.2.5",
